Renaming the tool to kubemci
