+++ README.md
@@ -1,6 +1,6 @@
-# mci
+# kubemci
 
-mci is a tool to configure kubernetes ingress to load balance traffic across
+kubemci is a tool to configure kubernetes ingress to load balance traffic across
 multiple kubernetes clustes.
 
 This is Work in Progress, not production ready yet.
