+++ app/kubemci/cmd/delete_test.go
