+++ app/kubemci/cmd/getstatus_test.go
