+++ app/kubemci/pkg/gcp/healthcheck/fake_healthchecksyncer.go
