+++ app/kubemci/pkg/gcp/healthcheck/interfaces.go
