+++ app/kubemci/pkg/gcp/namer/namer.go
