+++ app/kubemci/pkg/gcp/targetproxy/fake_targetproxysyncer.go
